requirements.txt
jsonschema==4.16.0
langchain==0.0.261
numpy==1.23.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
pandas==1.5.3
psutil==5.9.2
python-dotenv==1.0.0
streamlit==1.23.1
streamlit-card==0.0.61
streamlit-chat==0.1.1
streamlit-elements==0.1.0
streamlit-extras==0.3.0
matplotlib==3.7.1
